pyproject.toml
[tool.poetry]
name = "gallery"
version = "0.1.0.dev1"
description = "picture gallery"
authors = ["Julian Heise <jam.heise@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
Pillow = "^9.0.0"
exif = "^1.3.4"
psycopg2-binary = "^2.9.3"
fs = "^2.4.14"
click = "^8.0.3"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
ingest = "ingestion.cli:ingest"
